1. Change the logic of deleting chats so that now only their creators can delete chats; 2. Make the open chat delete button in an open chat invisible to users who are not the creators of this open chat; 3. Change the long name of the application in the file "manifest.json"; 4. Change automatically one file for "firebase" hosting.

+++ src/routes/ChatsList/ChangeChatsList/ChangeChatsList.js
@@ -146,11 +146,15 @@ export const ChangeChatsList = () => {
     if (valueName !== '') {
       const delChatsListRedKey = getKeyForTheChatByChatName(chatsListChatsKindOfDictRed, valueName);
       if (delChatsListRedKey) {
-        dispatch(removeFromChatsListWithThunkAction(delChatsListRedKey));
-        dispatch(removeAllMessagesInDeleteChatWithThunkAction(delChatsListRedKey));
-        const deleteChat = valueName;
-        setSuccess(`Чат "${deleteChat}" удален`);
-        resetValue();
+        if ((chatsListChatsKindOfDictRed[delChatsListRedKey]['chatAuthor'] && chatsListChatsKindOfDictRed[delChatsListRedKey]['chatAuthor'] === auth.currentUser.uid) || !chatsListChatsKindOfDictRed[delChatsListRedKey]['chatAuthor']) {
+          dispatch(removeFromChatsListWithThunkAction(delChatsListRedKey));
+          dispatch(removeAllMessagesInDeleteChatWithThunkAction(delChatsListRedKey));
+          const deleteChat = valueName;
+          setSuccess(`Чат "${deleteChat}" удален`);
+          resetValue();
+        } else {
+          setError('Чаты могут удалять только их авторы');
+        }
       } else {
         setError('Чат не найден');
       }
